isolate table into client component, keep page.tsx static
